requirements.txt
crewai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
streamlit
watchdog
pydantic
litellm
